# contracts/openapi/booking-service.v1.yaml
openapi: 3.1.0
info:
  title: SmartBus Booking Service API
  version: 2.0.0
  description: Booking lifecycle and orchestration contract for SmartBus ticket purchases.
servers:
  - url: http://localhost:8081
tags:
  - name: Booking
  - name: Admin
paths:
  /api/v1/bookings/foundation:
    get:
      operationId: getBookingFoundation
      summary: Get booking service foundation metadata
      tags: [Booking]
      responses:
        '200':
          description: Service metadata returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FoundationResponse'

  /api/v1/bookings/orchestrated-bookings:
    post:
      operationId: createOrchestratedBooking
      summary: Create a SmartBus booking through the orchestration flow
      tags: [Booking]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingRequest'
            examples:
              oneWayBooking:
                value:
                  customerName: Jane Rider
                  customerEmail: jane@example.com
                  fromStop: Downtown Terminal
                  toStop: Airport Station
                  tripDate: '2026-03-28'
                  tripType: one-way
                  passengers: 1
                  paymentMethodToken: tok_visa_4242
      responses:
        '201':
          description: Booking was accepted and completed.
          headers:
            Location:
              description: URL of the created booking resource.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '422':
          $ref: '#/components/responses/UnprocessableError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '504':
          $ref: '#/components/responses/GatewayTimeout'

  /api/v1/bookings/{bookingReference}:
    get:
      operationId: getBooking
      summary: Get full booking details by booking reference
      tags: [Booking]
      parameters:
        - $ref: '#/components/parameters/BookingReference'
      responses:
        '200':
          description: Booking returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingSummaryResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: cancelBooking
      summary: Cancel a booking by booking reference
      tags: [Booking]
      parameters:
        - $ref: '#/components/parameters/BookingReference'
      responses:
        '204':
          description: Booking cancelled successfully.
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v1/bookings/{bookingReference}/state:
    get:
      operationId: getBookingState
      summary: Get the persisted workflow state of a booking
      tags: [Booking]
      parameters:
        - $ref: '#/components/parameters/BookingReference'
      responses:
        '200':
          description: Booking state returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingProcessStateResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v1/bookings:
    get:
      operationId: listBookingsByCustomer
      summary: List bookings for a customer email
      tags: [Booking]
      parameters:
        - name: customerEmail
          in: query
          required: true
          schema:
            type: string
            format: email
      responses:
        '200':
          description: Bookings returned successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookingSummaryResponse'

  /api/v1/bookings/admin/bookings:
    get:
      operationId: listAllBookings
      summary: List all bookings (paginated)
      tags: [Admin]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Paginated booking list returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedBookingSummaryResponse'

components:
  parameters:
    BookingReference:
      name: bookingReference
      in: path
      required: true
      schema:
        type: string
      example: BK-1234ABCD

  responses:
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    ValidationError:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    UnprocessableError:
      description: Business rule rejected the request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    ServiceUnavailable:
      description: A downstream SmartBus service was unavailable after retries.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    GatewayTimeout:
      description: A downstream SmartBus service exceeded the configured timeout.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'

  schemas:
    FoundationResponse:
      type: object
      required: [service, responsibility, database, dependsOn, plannedOperations]
      properties:
        service:
          type: string
        responsibility:
          type: string
        database:
          type: string
        dependsOn:
          type: array
          items:
            type: string
        plannedOperations:
          type: array
          items:
            type: string

    BookingRequest:
      type: object
      required: [customerName, customerEmail, fromStop, toStop, tripDate, tripType, passengers, paymentMethodToken]
      properties:
        customerName:
          type: string
          minLength: 1
        customerEmail:
          type: string
          format: email
        fromStop:
          type: string
          minLength: 1
        toStop:
          type: string
          minLength: 1
        tripDate:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}$'
        tripType:
          type: string
          enum: [one-way, round-trip]
        passengers:
          type: integer
          minimum: 1
          maximum: 6
        paymentMethodToken:
          type: string
          minLength: 1

    BookingResponse:
      type: object
      required: [bookingReference, correlationKey, status, routeCode, paymentTransactionId, notificationId, totalAmount, workflowLog]
      properties:
        bookingReference:
          type: string
        correlationKey:
          type: string
        status:
          type: string
        routeCode:
          type: string
        paymentTransactionId:
          type: string
        notificationId:
          type: string
        totalAmount:
          type: number
          format: double
        workflowLog:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowLogEntry'

    BookingSummaryResponse:
      type: object
      required: [bookingReference, customerName, customerEmail, fromStop, toStop, tripDate, tripType, passengers, currentState]
      properties:
        bookingReference:
          type: string
        customerName:
          type: string
        customerEmail:
          type: string
        fromStop:
          type: string
        toStop:
          type: string
        tripDate:
          type: string
        tripType:
          type: string
        passengers:
          type: integer
        routeCode:
          type: string
          nullable: true
        departureTime:
          type: string
          nullable: true
        arrivalTime:
          type: string
          nullable: true
        totalAmount:
          type: number
          format: double
        paymentTransactionId:
          type: string
          nullable: true
        notificationId:
          type: string
          nullable: true
        currentState:
          type: string
        lastError:
          type: string
          nullable: true

    PagedBookingSummaryResponse:
      type: object
      required: [items, page, size, totalElements, totalPages]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/BookingSummaryResponse'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer

    BookingProcessStateResponse:
      type: object
      required: [bookingReference, correlationKey, currentState]
      properties:
        bookingReference:
          type: string
        correlationKey:
          type: string
        currentState:
          type: string
        lastError:
          type: string
          nullable: true

    WorkflowLogEntry:
      type: object
      required: [step, element, detail]
      properties:
        step:
          type: string
        element:
          type: string
          enum: [receive, invoke, flow, switch, reply]
        detail:
          type: string

    ApiErrorResponse:
      type: object
      required: [status, code, message, details, path]
      properties:
        status:
          type: integer
          example: 400
        code:
          type: string
          example: VALIDATION_FAILED
        message:
          type: string
        details:
          type: array
          items:
            type: string
        path:
          type: string